% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{checkHeight}
\alias{checkHeight}
\title{checkHeight}
\usage{
checkHeight(students.input, sex.specific = TRUE, print.statement = FALSE)
}
\arguments{
\item{students.input}{a input dataframe}

\item{sex.specific}{logical value}

\item{print.statement}{logical value}

\item{r}{a vector}
}
\value{
result of the computation
}
\description{
Script for simple function that checks the difference in height     from the sex-specific mean or the overall mean height for each of the students
}
\details{
an example function named 'checkheight'
Script for simple function that checks the difference in height from the sex-
specific mean or the overall mean height for each of the students in the given #'dataframe
Date: 02.11.2017
}
\author{
Kunlan Li <li.kunlan@campus.lmu.de>
}
